backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-multipart==0.0.6
PyMuPDF==1.25.3
httpx==0.27.0
sqlalchemy==2.0.25
chromadb==0.6.3
sentence-transformers==3.0.1
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pytest==8.0.0
pytest-asyncio==0.23.3
